Starting to refactor: Separating frontend stuff into netplay_frontend.c
